refactor: extract generation set into module
