+++ src/lib.rs
@@ -1,3 +1,5 @@
 mod generation;
+mod generation_set;
 
-pub use generation::{Generation, GenerationSet};
+pub use generation::Generation;
+pub use generation_set::GenerationSet;
